+++ src/components/Bundlr.tsx
@@ -10,25 +10,22 @@ import {
   Text,
   VStack,
   HStack,
-  // Box,
 } from '@chakra-ui/react';
-// import { Alert } from 'components/Alert';
 import { useBundlr } from 'features/embalm/stepContent/hooks/useBundlr';
 import { useUploadPrice } from 'features/embalm/stepNavigator/hooks/useUploadPrice';
 import { useCallback, useState } from 'react';
 import { setBalance } from 'store/bundlr/actions';
-import { useDispatch, useSelector } from 'store/index';
+import { useDispatch } from 'store/index';
 import { useBundlrSession } from '../features/embalm/stepContent/hooks/useBundlrSession';
 import { useGetBalance } from '../features/embalm/stepContent/hooks/useGetBalance';
 import { uploadPriceDecimals } from 'lib/constants';
 
-export function Bundlr() {
+export function Bundlr({ children }: { children?: React.ReactNode }) {
   const dispatch = useDispatch();
   const { fund, isFunding } = useBundlr();
   const { connectToBundlr, isConnected, disconnectFromBundlr } = useBundlrSession();
-  const { balance, getBalance, formattedBalance } = useGetBalance();
+  const { getBalance, formattedBalance } = useGetBalance();
   const { uploadPrice, formattedUploadPrice } = useUploadPrice();
-  // const file = useSelector(x => x.embalmState.file);
   const [amount, setAmount] = useState(parseFloat(uploadPrice || '0').toFixed(uploadPriceDecimals));
   const ONE_WEI = '0.000000000000000001';
 
@@ -49,22 +46,6 @@ export function Bundlr() {
     const newBalance = await getBalance();
     dispatch(setBalance(newBalance));
   }, [amount, dispatch, fund, getBalance, isAmountValid]);
-
-  // function AlertMessage() {
-  //   if (!file) {
-  //     return <Alert status="warning">Upload a payload to get the upload price.</Alert>;
-  //   }
-
-  //   if (parseFloat(balance) < parseFloat(uploadPrice)) {
-  //     return <Alert status="warning">You need to add funds to Bundlr.</Alert>;
-  //   }
-
-  //   return (
-  //     <Alert status="success">
-  //       You have enough funds in Bundlr. You can top up your balance or continue to the next step.
-  //     </Alert>
-  //   );
-  // }
 
   return (
     <VStack
@@ -111,9 +92,7 @@ export function Bundlr() {
               </Button>
             </HStack>
           </FormControl>
-          {/* <Box py={3}>
-            <AlertMessage />
-          </Box> */}
+          {children}
           <VStack
             border="1px solid "
             borderColor="violet.700"
